@@ -1,16 +1,16 @@
 name: flutter_progress_dialog
 description: A flutter library for the progress dialog. Easily show and hide. Support specify custom loading widget.
-version: 0.1.0
+version: 0.1.1
 author: Dylan Wu <dylanwuzh@gmail.com>
 homepage: https://github.com/wuzhendev/flutter_progress_dialog
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_platform_widgets: ^1.9.0
+  flutter_platform_widgets: ^3.0.0
 
 dev_dependencies:
   flutter_test:
